requirements.txt
requests==2.14.2
Django==1.11.2
pandas==0.20.1
booleano==1.1a1
Shapely==1.6b4
PyYAML==3.12
django-favicon==0.1.3